tcp: define macros for a couple reclaim thresholds
Git-commit: 5823fc96d754d824bb298622055e8dd5e1252122
Message-Id: <5823fc96d754d824bb298622055e8dd5e1252122.1637135496.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.16-rc1

A following patch is going to implement a similar reclaim schema
for the MPTCP protocol, with different locking.

Let's define a couple of macros for the used thresholds, so
that the latter code will be more easily maintainable.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.16-rc1-tcp-define-macros-for-a-couple-reclaim-thresholds.patch