scripts: introduce patches
This python script is used to create a list which records all the
mainline commits (those are merged in SUSE Euler kernel). We can
just switch to SUSE-Euler-{1,2}.0 or relevant branch, go to the
kernel-source directory and call

$ /path/to/source/of/kernel-source/patches > /tmp/suse-patches.list

Then we can run git-fixes against the list to get needed fixes.

Note, this file is imported from Joerg's github tree, thanks for
his awesome work of git-fixes!

[1]. https://github.com/joergroedel/git-fixes

Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Change-Id: I0d2652016443e1e94f8ca2e3cda4507682fca099